Break up giant file
